ml-model/requirements.txt
async-timeout==5.0.1
audioread==3.0.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
fonttools==4.55.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
matplotlib==3.9.3
msgpack==1.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
soundfile==0.13.1
soxr==0.5.0.post1
spotipy==2.24.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
